client_cli/pytests/pyproject.toml
[tool.poetry]
name = "pytestclientcli"
version = "0.1.0"
description = ""
authors = ["alexstroke <busyfifer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faker = "*"
allure-python-commons = "*"
cryptography = "*"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
allure-pytest = "*"
pytest-xdist = "*"
mypy = "*"
isort = "*"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



